\name{Doench_2016_processing}
\alias{Doench_2016_processing}
\title{Donech 2016 Processsing
}
\description{Internal function that assigns an effciency score based on the 2016 Donech Efficiency Scoring Method
}
\usage{
Doench_2016_processing(seqlist)
}
\arguments{
  \item{seqlist}{
    A list of sgRNA (30-mer).
  }
}
\author{
Dylan Beeber
}