\name{getofftargetdata}
\alias{getofftargetdata}
\title{Off Target Data Frame Creation}
\description{
Will provide a data frame with all information about the generated sgRNA returned by the sgRNA_design function.
}
\usage{
getofftargetdata(x)
}
\arguments{
  \item{x}{the data list generated by the sgRNA_design function
  }
}
\author{
Dylan Beeber
}
\examples{

## First generate data with the sgRNA_Design Function:
testseq <- "GGCAGAGCTTCGTATGTCGGCGATTCATCTCAAGTAGAAGATCCTGGTGCAGTAGG"
usergenome <- BSgenome.Scerevisiae.UCSC.sacCer2::BSgenome.Scerevisiae.UCSC.sacCer2
gtfname <- "Saccharomyces_cerevisiae.R64-1-1.92.gtf.gz"
annotation_file <- system.file("example_data", gtfname, package = "crispRdesignR")
alldata <- sgRNA_design(testseq, usergenome, annotation_file)

## Then separate and format the off-target data with getofftargetdata()
final_data <- getofftargetdata(alldata)
}
